requirements.txt
ctransformers
sentence-transformers
pinecone-client
langchain
flask
pypdf
python-dotenv
langchain-community
-e .
